Turn on more warnings in ios / mac projects. Make casts explicit.   This makes casts that loose precision explicit, from here on we will get warnings.   The changes in this commit are made without evaluating each cast, asuming the original casts were intentional.
Patch by: jakerr@google.com
Review: https://breakpad.appspot.com/453002/



git-svn-id: http://google-breakpad.googlecode.com/svn/trunk@1046 4c0a9323-5329-0410-9bdc-e9ce6186880e

+++ src/client/ios/BreakpadController.mm
@@ -224,7 +224,7 @@ NSString* GetPlatform() {
 
   if (spanSeconds >= uploadIntervalInSeconds_)
     return 0;
-  return uploadIntervalInSeconds_ - spanSeconds;
+  return uploadIntervalInSeconds_ - static_cast<int>(spanSeconds);
 }
 
 - (void)reportWillBeSent {
